Add files via upload
Added the ideas discussed in the past meeting. Still need to finish out the individual form results, right now they're all the same, but the display works. Also, need to add more of the locations. And need to make aesthetic changes to make this section congruent with the rest of the app.